Restructured the folders. Refactored the code to just be fully async and if not possible to also be waited for. Added optional botname and avatarurl to be configured. Rewritten the readme to also include a apssettings.json part. Updated Serilog and Discord.Webhook used to be the newest available version. Added comments for future developers to now what is doing what.
